Set up register page
