+++ src/css/colors.scss
@@ -1,4 +1,5 @@
 * {
+  --primary: #7000cc;
   --gray: #6d747c;
   --red: #cc444a;
 }
