+++ src/css/forms/buttons.scss
@@ -5,4 +5,17 @@
   border-radius: 3px;
   margin: 10px;
   font-weight: 700;
+  border: none;
+  box-shadow: 1px 2px #999;
+
+  &:active {
+    box-shadow: none;
+    transform: translateY(1px);
+  }
+}
+
+.btn-primary {
+  background-color: var(--primary);
+  box-shadow: 1px 2px #44007b;
+  color: #fff;
 }
