+++ src/css/index.scss
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css?family=Oswald:300,400,700|Teko:400,700');
+
 @import './colors.scss';
 @import './utils.scss';
 
@@ -8,14 +10,27 @@
 @import './components/loading.scss';
 
 @import './app.scss';
+@import './views/register.scss';
+
+* {
+  box-sizing: border-box;
+}
 
 body {
   margin: 0;
   padding: 0;
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
-    'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  font-family: 'Oswald', 'Roboto', 'Helvetica Neue', sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+}
+
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+  font-family: 'Teko', 'Oswald', 'Roboto', 'Helvetica Neue', sans-serif;
 }
 
 code {
